feat: cutouts can use pts or bboxes
